[Feat]: 선택한 환율로 통화를 변경이 가능하게 actionSheet 추가

+++ TripLog/CoreData/DataClass/SubCard+CoreDataProperties.swift
@@ -19,7 +19,7 @@ extension SubCard {
 
     @NSManaged public var id: UUID?
     @NSManaged public var imagesData: NSObject?
-    @NSManaged public var price: Int64
+    @NSManaged public var price: Double
     @NSManaged public var script: String?
     @NSManaged public var starsState: [Bool]?
     @NSManaged public var title: String?
